refactor(all): move files from src/config to app workspace
