+++ app/configs/module.config.d.ts
@@ -1,5 +1,3 @@
-/// <reference types="vite/client" />
-
 declare module '*.vue' {
   import type { DefineComponent } from 'vue'
   const component: DefineComponent<
@@ -8,10 +6,4 @@ declare module '*.vue' {
     unknown
   >
   export default component
-}
-
-declare module '*.md' {
-  import type { ComponentOptions } from 'vue'
-  const Component: ComponentOptions
-  export default Component
 }
