+++ app/configs/path.config.ts
